Next.js 15대응 inquiry 페이지에 Suspense적용
